% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{vnote32}
\alias{vnote32}
\title{Transform VNoteX 3.x notebook to VNote 2.x notebook}
\usage{
vnote32(from, to = tempdir(), mode = c("replace", "coexist"))
}
\arguments{
\item{from}{is the notebook directory of 3.x}

\item{to}{is the new notebook directory of 2.x}

\item{mode}{replace (default), or coexist}
}
\description{
Transform VNoteX 3.x notebook to VNote 2.x notebook
}
\examples{
\dontrun{
library(vnoteTransformer)
vnote32(from = "E:/VNoteBook/", to = "E:/VNotebook2/")
}
}
